Default country and province upon address update
