+++ source/frontend/src/features/mapSideBar/tabs/propertyDetails/update/UpdatePropertyDetailsForm.tsx
@@ -9,7 +9,7 @@ import * as API from 'constants/API';
 import { PropertyAdjacentLandTypes, PropertyTenureTypes } from 'constants/index';
 import { FormikProps, getIn, useFormikContext } from 'formik';
 import { useLookupCodeHelpers } from 'hooks/useLookupCodeHelpers';
-import React, { useEffect, useMemo } from 'react';
+import React, { useEffect } from 'react';
 import { Col, Row } from 'react-bootstrap';
 import styled from 'styled-components';
 import { prettyFormatDate } from 'utils';
@@ -54,15 +54,6 @@ export const UpdatePropertyDetailsForm: React.FunctionComponent<
   const regionOptions = getOptionsByType(API.REGION_TYPES);
   const districtOptions = getOptionsByType(API.DISTRICT_TYPES);
 
-  const countryCA = useMemo(
-    () => getByType(API.COUNTRY_TYPES).find(c => c.code === 'CA'),
-    [getByType],
-  );
-  const provinceBC = useMemo(
-    () => getByType(API.PROVINCE_TYPES).find(p => p.code === 'BC'),
-    [getByType],
-  );
-
   // multi-selects
   const tenureStatus = getIn(values, 'tenures') as PropertyTenureFormModel[];
   const adjacentLands = getIn(values, 'adjacentLands') as PropertyAdjacentLandFormModel[];
@@ -83,12 +74,6 @@ export const UpdatePropertyDetailsForm: React.FunctionComponent<
   const volumetricUnit = getIn(values, 'volumetricUnitTypeCode') as string;
 
   const setFieldValue = formikProps.setFieldValue;
-
-  // property addresses are limited to BC, Canada
-  useEffect(() => {
-    setFieldValue('address.province', provinceBC, false);
-    setFieldValue('address.country', countryCA, false);
-  }, [countryCA, provinceBC, setFieldValue]);
 
   // clear related fields when volumetric parcel radio changes
   useEffect(() => {
